[refactor] Make directories separate for servers and clients
