--- src/app/_libs/supabase/client-session-provider.tsx
@@ -1,11 +1,0 @@
-'use client';
-
-import { SessionContextProvider } from '@supabase/auth-helpers-react';
-
-import { supabase } from './client-instance';
-
-import type { FC, PropsWithChildren } from 'react';
-
-export const SupabaseClientSessionProvider: FC<PropsWithChildren> = ({ children }) => {
-  return <SessionContextProvider supabaseClient={supabase}>{children}</SessionContextProvider>;
-};
